+++ NEWS.md
@@ -1,6 +1,31 @@
-# xnet 0.1.4
+## xnet 0.1.5
 
-## breaking changes
+### Breaking changes
+
+* class `tskrr`, `tskrrHeterogenous` and `tskrrHomogenous`:
+    - the slot `has.orig` has been removed as it doesn't make sense to 
+ keep the original kernel matrices. It is replaced by a slot `has.hat`
+ allowing to store the hat matrices.
+    - the slots `k.orig` and `g.orig` have been replaced by the slots
+ `Hk` and `Hg` to store the hat matrices. These are more needed for
+ fitting etc. 
+ 
+* The function `has_original` has been removed and replaced by `has_hat`
+* The argument `keep` of the function `tskrr` now stores the hat matrices
+instead of the original kernel matrices.
+* The function `tskrr` has lost its argument `homogenous`. It didn't make
+sense to set that by hand.
+
+### New features
+
+* classes `tskrrHeterogenousImpute` and `tskrrHomogenousImpute` are added
+  to allow for storing models with imputed predictions.
+
+### bug fixes and minor improvements
+
+## xnet 0.1.4
+
+### breaking changes
 
 * `get_loo_fun()` : 
 
@@ -9,7 +34,7 @@ allows for extension of the function based on either an object or
 the class of that object. 
     - `x` becomes the first argument.
 
-## New features
+### New features
 
 * There's a new function `linear_filter` that fits a linear filter over
 an adjacency matrix. This function comes with a class `linearFilter`.
@@ -18,7 +43,7 @@ for optimization.
 * functions `loss_mse()` and `loss_auc()` are provided for tuning.
 * `update()` allows to retrain the model with new lambdas.
 
-## bug fixes and minor improvements
+### bug fixes and minor improvements
 
 * predictions were calculated wrongly in `tune()`: fixed.
 * MSE was calculated wrongly in the previous version of `tune()`: fixed.